Fix implicit int64_t -> uint64_t conversions trip -Wsign-conversion

+++ include/valijson/schema_parser.hpp
@@ -1993,7 +1993,7 @@ private:
             const int64_t value = node.asInteger();
             if (value >= 0) {
                 constraints::MaxItemsConstraint constraint;
-                constraint.setMaxItems(value);
+                constraint.setMaxItems(static_cast<uint64_t>(value));
                 return constraint;
             }
         }
@@ -2017,7 +2017,7 @@ private:
             const int64_t value = node.asInteger();
             if (value >= 0) {
                 constraints::MaxLengthConstraint constraint;
-                constraint.setMaxLength(value);
+                constraint.setMaxLength(static_cast<uint64_t>(value));
                 return constraint;
             }
         }
@@ -2043,7 +2043,7 @@ private:
             int64_t value = node.asInteger();
             if (value >= 0) {
                 constraints::MaxPropertiesConstraint constraint;
-                constraint.setMaxProperties(value);
+                constraint.setMaxProperties(static_cast<uint64_t>(value));
                 return constraint;
             }
         }
@@ -2121,7 +2121,7 @@ private:
             const int64_t value = node.asInteger();
             if (value >= 0) {
                 constraints::MinItemsConstraint constraint;
-                constraint.setMinItems(value);
+                constraint.setMinItems(static_cast<uint64_t>(value));
                 return constraint;
             }
         }
@@ -2144,7 +2144,7 @@ private:
             const int64_t value = node.asInteger();
             if (value >= 0) {
                 constraints::MinLengthConstraint constraint;
-                constraint.setMinLength(value);
+                constraint.setMinLength(static_cast<uint64_t>(value));
                 return constraint;
             }
         }
@@ -2170,7 +2170,7 @@ private:
             int64_t value = node.asInteger();
             if (value >= 0) {
                 constraints::MinPropertiesConstraint constraint;
-                constraint.setMinProperties(value);
+                constraint.setMinProperties(static_cast<uint64_t>(value));
                 return constraint;
             }
         }
